Include Gemfile that references our gemspec so we can simply do `bundle` to get all dependencies
Having a Gemfile would allow us to easily retrieve project dependencies by doing `bundle`.  Since we have a gemspec in place, we can simply refer to it so we do not duplicate the dependencies in the gemspec and in Gemfile.

One catch though is that the `gemspec` directive in Gemfile expects a .gemspec file, so we have to rename osaka.gemfile to osaka.gemspec.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>