Replace "Don't Save" button to "Delete" as Mountain Lion uses the latter
Mountain Lion introduced deeper iCloud integration to its iWork apps as well
as core OS X apps such as TextEdit and Preview.  We now get an additional
sheet wherein we can save files directly into iCloud.  Part of this change is
that the "Don't Save" button was replaced with "Delete" instead.

Since Osaka relies on the text "Don't Save" certain specs in Mountain Lion
fail.  Replacing them to "Delete" makes the test pass but I would think it would
break the specs in Lion and earlier.  We may need to introduce some sort of
intelligence to determine which button text exists based on the OS.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>